Change the way to determine if a motion sensor can turn on a switch.
Previously this value is associated with the switch itself. However, the
zone that contains the switch might have multiple motion sensors or
motion sensor equivalent. One of them might disable turning on the
switch, but the rest might not. Thus it makes more sense to have this
attribute associated with the individual motion sensor.

This allows turning on a zone's light when the TV is on (here the TV
power status acts as a motion sensor), even if the zone shares another
motion sensor with an adjacent zone.