Use lib feedparser to retrieve alert RSS feed (instead of OH feed add-on)
The OH add-on has a limitation: cannot handle multiple entries in the
feed update. Recently Environment Canada has added a notice of future
feed content as the first entry of the feed (the actual weather alert
entry is the second). As a result, weather alert hasn't been working for
a while. With this change, the OH feed add-on can be removed.